Standardize names some.
