--- src/pages/Grid.scss
@@ -1,38 +1,0 @@
-$testColor: #333333;
-
-.gridPage {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    
-}
-
-.activeTests {
-    width: 200px;   
-}
-
-.ghostgrid {
-
-    border-spacing: 0px;
-    border-collapse: collapse;
-
-    .test{
-        background-color: $testColor;
-        color: #FFF;
-        padding: 5px 10px;
-        border-radius: 10px;
-
-        cursor: pointer;
-        
-        &:hover{
-            background-color: lighten($testColor, 10%);
-        }
-    }
-    .ghost{
-
-    }
-
-    td {
-        border: solid 1px black;
-    }
-}
